# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wavecrest Pro API",
    "version": "1.0.0",
    "description": "Wavecrest Pro is a real-time trend intelligence platform that aggregates trending topics across multiple social platforms. This API provides endpoints for authentication, trend discovery, dashboard management, billing, and analytics.",
    "contact": {
      "name": "Wavecrest Pro Support",
      "url": "https://wavecrest.pro",
      "email": "support@wavecrest.pro"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://wavecrest.pro",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/stripe/create-checkout": {
      "post": {
        "tags": ["Billing"],
        "summary": "Create a Stripe checkout session",
        "description": "Initiate a Stripe Checkout session for subscribing to a paid plan. Returns a URL to redirect the user to Stripe's hosted checkout page.",
        "operationId": "createCheckout",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["plan"],
                "properties": {
                  "plan": {
                    "type": "string",
                    "enum": ["plus", "pro", "max", "teams", "enterprise"],
                    "description": "Plan ID to subscribe to"
                  }
                }
              },
              "example": {
                "plan": "pro"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "description": "Stripe Checkout URL to redirect user to"
                    }
                  }
                },
                "example": {
                  "url": "https://checkout.stripe.com/c/pay/cs_live_abc123..."
                }
              }
            }
          },
          "400": {
            "description": "Invalid plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid plan selected"
                }
              }
            }
          }
        }
      }
    },
    "/api/stripe/portal": {
      "post": {
        "tags": ["Billing"],
        "summary": "Open Stripe billing portal",
        "description": "Create a Stripe Customer Portal session. The portal allows users to manage their subscription, update payment methods, and view invoices.",
        "operationId": "createPortal",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Portal session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "description": "Stripe Customer Portal URL"
                    }
                  }
                },
                "example": {
                  "url": "https://billing.stripe.com/p/session/abc123..."
                }
              }
            }
          },
          "400": {
            "description": "No active subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "No active subscription found"
                }
              }
            }
          }
        }
      }
    },
    "/api/stripe/cancel": {
      "post": {
        "tags": ["Billing"],
        "summary": "Cancel subscription",
        "description": "Cancel the current subscription. The subscription remains active until the end of the current billing period, then downgrades to the Free plan.",
        "operationId": "cancelSubscription",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Subscription cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "effective_date": { "type": "string", "format": "date-time" }
                  }
                },
                "example": {
                  "message": "Subscription cancelled. Active until end of billing period.",
                  "effective_date": "2025-02-15T00:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "No active subscription to cancel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "No active subscription to cancel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after login or signup"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" }
        }
      },
...
    }
  }
}